Integration of trees into the main scene which is now at night and various refactors

+++ cameraController.c
@@ -38,9 +38,9 @@ void updateCamera(float xoffset, float yoffset) {
 void defaultCameraTransforms(vec3 *pos, vec3 *dir, float distance, vec2 angles) {
 	float yaw = angles.x;
 	float pitch = angles.y;
-	float x = cos(yaw) * cos(pitch);
-	float y = sin(pitch);
-	float z = sin(yaw) * cos(pitch);
+	float x = cosf(yaw) * cosf(pitch);
+	float y = sinf(pitch);
+	float z = sinf(yaw) * cosf(pitch);
 
 	vec3 direction = (vec3){x, y, z};
 	vec3 position = vec3_scale(direction, -distance);
